[refactor&fix]raw password

+++ src/main/java/github/javaguide/springsecurityjwtguide/system/service/UserService.java
@@ -1,17 +1,18 @@
 package github.javaguide.springsecurityjwtguide.system.service;
 
 import com.google.common.collect.ImmutableMap;
-import github.javaguide.springsecurityjwtguide.system.web.representation.UserRepresentation;
-import github.javaguide.springsecurityjwtguide.system.web.request.UserRegisterRequest;
 import github.javaguide.springsecurityjwtguide.system.entity.Role;
 import github.javaguide.springsecurityjwtguide.system.entity.User;
 import github.javaguide.springsecurityjwtguide.system.entity.UserRole;
 import github.javaguide.springsecurityjwtguide.system.enums.RoleType;
-import github.javaguide.springsecurityjwtguide.system.exception.ResourceNotFoundException;
+import github.javaguide.springsecurityjwtguide.system.exception.RoleNotFoundException;
 import github.javaguide.springsecurityjwtguide.system.exception.UserNameAlreadyExistException;
+import github.javaguide.springsecurityjwtguide.system.exception.UserNameNotFoundException;
 import github.javaguide.springsecurityjwtguide.system.repository.RoleRepository;
 import github.javaguide.springsecurityjwtguide.system.repository.UserRepository;
 import github.javaguide.springsecurityjwtguide.system.repository.UserRoleRepository;
+import github.javaguide.springsecurityjwtguide.system.web.representation.UserRepresentation;
+import github.javaguide.springsecurityjwtguide.system.web.request.UserRegisterRequest;
 import github.javaguide.springsecurityjwtguide.system.web.request.UserUpdateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,6 +22,8 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Objects;
+
 /**
  * @author shuang.kou
  */
@@ -28,7 +31,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class UserService {
 
-    public static final String USERNAME = "username";
+    public static final String USERNAME = "username:";
     private final UserRepository userRepository;
     private final RoleRepository roleRepository;
     private final UserRoleRepository userRoleRepository;
@@ -36,29 +39,40 @@ public class UserService {
 
     @Transactional(rollbackFor = Exception.class)
     public void save(UserRegisterRequest userRegisterRequest) {
-        checkUserNameNotExist(userRegisterRequest.getUserName());
-        User user = User.of(userRegisterRequest);
+        ensureUserNameNotExist(userRegisterRequest.getUserName());
+        User user = userRegisterRequest.toUser();
         user.setPassword(bCryptPasswordEncoder.encode(userRegisterRequest.getPassword()));
         userRepository.save(user);
         //给用户绑定两个角色：用户和管理者
-        Role studentRole = roleRepository.findByName(RoleType.USER.getName()).orElseThrow(() -> new ResourceNotFoundException(ImmutableMap.of("roleName", RoleType.USER.getName())));
-        Role managerRole = roleRepository.findByName(RoleType.MANAGER.getName()).orElseThrow(() -> new ResourceNotFoundException(ImmutableMap.of("roleName", RoleType.MANAGER.getName())));
+        Role studentRole = roleRepository.findByName(RoleType.USER.getName()).orElseThrow(() -> new RoleNotFoundException(ImmutableMap.of("roleName", RoleType.USER.getName())));
+        Role managerRole = roleRepository.findByName(RoleType.MANAGER.getName()).orElseThrow(() -> new RoleNotFoundException(ImmutableMap.of("roleName", RoleType.MANAGER.getName())));
         userRoleRepository.save(new UserRole(user, studentRole));
         userRoleRepository.save(new UserRole(user, managerRole));
     }
 
     public User find(String userName) {
-        return userRepository.findByUserName(userName).orElseThrow(() -> new ResourceNotFoundException(ImmutableMap.of(USERNAME, userName)));
+        return userRepository.findByUserName(userName).orElseThrow(() -> new UserNameNotFoundException(ImmutableMap.of(USERNAME, userName)));
     }
 
     public void update(UserUpdateRequest userUpdateRequest) {
-        User user = userRepository.findByUserName(userUpdateRequest.getUserName()).orElseThrow(() -> new ResourceNotFoundException(ImmutableMap.of(USERNAME, userUpdateRequest.getUserName())));
-        user.updateFrom(userUpdateRequest);
+        User user = find(userUpdateRequest.getUserName());
+        if (Objects.nonNull(userUpdateRequest.getFullName())) {
+            user.setFullName(userUpdateRequest.getFullName());
+        }
+        if (Objects.nonNull(userUpdateRequest.getPassword())) {
+            user.setPassword(bCryptPasswordEncoder.encode(userUpdateRequest.getPassword()));
+        }
+        if (Objects.nonNull(userUpdateRequest.getEnabled())) {
+            user.setEnabled(userUpdateRequest.getEnabled());
+        }
         userRepository.save(user);
     }
 
 
     public void delete(String userName) {
+        if (!userRepository.existsByUserName(userName)) {
+            throw new UserNameNotFoundException(ImmutableMap.of(USERNAME, userName));
+        }
         userRepository.deleteByUserName(userName);
     }
 
@@ -66,7 +80,7 @@ public class UserService {
         return userRepository.findAll(PageRequest.of(pageNum, pageSize)).map(User::toUserRepresentation);
     }
 
-    private void checkUserNameNotExist(String userName) {
+    private void ensureUserNameNotExist(String userName) {
         boolean exist = userRepository.findByUserName(userName).isPresent();
         if (exist) {
             throw new UserNameAlreadyExistException(ImmutableMap.of(USERNAME, userName));
